+++ NEWS.md
@@ -8,7 +8,7 @@ Enhancements:
  - Return multiple turbo-carto errors #541.
 
 Announcements:
- - Upgrades turbo-carto to [0.13.0](https://github.com/CartoDB/turbo-carto/releases/tag/0.13.0).
+ - Upgrades turbo-carto to [0.14.0](https://github.com/CartoDB/turbo-carto/releases/tag/0.14.0).
 
 
 ## 2.60.0